rewrite the durations stuff. Next steps are to 1. Unify Duration and Duration block potentially just embedding the Heartbeat. 2. Fix tests 3. Heavy refactor of the workflow. Seems there are still a couple of descrepancies. We've also made terrible "imprompty" decisions about various program entities that needs to be updated
